- Abstracted out the sound player - Added sound player to test suite and NutManager - Refactored GetNotificationForMessage() method down to one line
