+++ SquirrelFinder.Tests/TheSquirrelFinder.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Net;
 using SquirrelFinder.Nuts;
 using System.Threading.Tasks;
+using SquirrelFinder.Sounds;
 
 namespace SquirrelFinder.Tests
 {
@@ -14,12 +15,14 @@ namespace SquirrelFinder.Tests
     public class TheSquirrelFinder
     {
         static NutManager _monitor;
+        static SoundManager _soundManager;
         static INut _nut;
 
         [TestInitialize]
         public void Setup()
         {
             _monitor = new NutManager();
+            _soundManager = new SoundManager();
             _nut = new Nut(new Uri("http://localhost"));
         }
 
@@ -82,12 +85,11 @@ namespace SquirrelFinder.Tests
         [TestMethod]
         public void CanMakeASound()
         {
-            _monitor.PlayTone(SquirrelFinderSound.FlatLine);
-            _monitor.PlayTone(SquirrelFinderSound.Gears);
-            _monitor.PlayTone(SquirrelFinderSound.Squirrel);
-            _monitor.PlayTone(SquirrelFinderSound.None);
-
-            _monitor.PlayTone(SquirrelFinderSound.None);
+            _soundManager.PlayTone(SquirrelFinderSound.FlatLine);
+            _soundManager.PlayTone(SquirrelFinderSound.Gears);
+            _soundManager.PlayTone(SquirrelFinderSound.Squirrel);
+            _soundManager.PlayTone(SquirrelFinderSound.None);
+            _soundManager.PlayTone(SquirrelFinderSound.None);
         }
 
         [TestClass]
