+++ SquirrelFinder/NutManager.cs
@@ -4,27 +4,14 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Media;
 using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace SquirrelFinder
 {
-
-    public enum SquirrelFinderSound
-    {
-        None,
-        FlatLine,
-        Gears,
-        Squirrel
-    }
-
     public class NutManager : INutManager
     {
-        static SoundPlayer _player = new SoundPlayer();
-        SquirrelFinderSound _currentSound = SquirrelFinderSound.None;
-
         public event EventHandler<NutCollectionEventArgs> NutCollectionChanged;
         public event EventHandler<NutEventArgs> NutChanged;
         public virtual void OnNutCollectionChanged(NutCollectionEventArgs e)
@@ -40,10 +27,6 @@ namespace SquirrelFinder
         FileSystemWatcher watcher = new FileSystemWatcher();
 
         public NutState CurrentState = NutState.Found;
-
-        static string _squirrelSound = AppDomain.CurrentDomain.BaseDirectory + "\\sounds\\squirrel.wav";
-        static string _gearsSound = AppDomain.CurrentDomain.BaseDirectory + "\\sounds\\gears.wav";
-        static string _flatLine = AppDomain.CurrentDomain.BaseDirectory + "\\sounds\\flatline.wav";
 
         private List<INut> _nuts;
         public List<INut> Nuts { get { return _nuts; } set { _nuts = value; } }
@@ -100,36 +83,6 @@ namespace SquirrelFinder
                 {
                     nut.Peek();
                 });
-            }
-        }
-
-        public virtual void PlayTone(SquirrelFinderSound sound)
-        {
-            if (sound == _currentSound) return;
-            _currentSound = sound;
-
-            var soundToPlay = "";
-            switch (sound)
-            {
-                case SquirrelFinderSound.FlatLine:
-                    soundToPlay = _flatLine;
-                    CurrentState = NutState.Lost;
-                    break;
-                case SquirrelFinderSound.Gears:
-                    soundToPlay = _gearsSound;
-                    CurrentState = NutState.Searching;
-                    break;
-                case SquirrelFinderSound.Squirrel:
-                    soundToPlay = _squirrelSound;
-                    CurrentState = NutState.Found;
-                    break;
-            }
-
-            if (sound != SquirrelFinderSound.None)
-            {
-                _player.Stop();
-                _player.SoundLocation = soundToPlay;
-                _player.Play();
             }
         }
     }
