+++ SquirrelFinder/SquirrelFinder.csproj
@@ -64,6 +64,7 @@
     <Compile Include="NutWatcher.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="NutSaver.cs" />
+    <Compile Include="Sounds\SoundManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
